Move transactions service to the shared area
