+++ src/test/transactions.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { TransactionsService } from '../app/shared/transactions.service';
+import { TransactionsService } from '../app/shared/services/transactions.service';
 
 describe('TransAccountService', () => {
   let service: TransactionsService;
